Created events service interface
