--- src/standard-callback.ts
@@ -1,4 +1,0 @@
-export class StandardCallback {
-    success: Function;
-    fail: Function;
-}
