+++ tools/client-mock-servlet.html
@@ -7,7 +7,7 @@
     <!--script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script-->
     <script src="indexmockserver.bundle.js"></script>
     <script>
-        //import {MockTracerServlet} from "./mock-tracer-servlet";
+        //import {EventsServiceSimpleMock} from "./mock-tracer-servlet";
 
         var tracer;
         window.onload = function(){
